+++ test/utility/test_api.py
@@ -1,33 +1,24 @@
+import datetime
 import os
 
 import pytest
 import rasterio
-import responses
 from rasterio.coords import BoundingBox
 from requests import Response
 from responses import matchers
 
-from climatoology.utility.api import LulcUtility, LulcWorkUnit
 from climatoology.utility.exception import PlatformUtilityException
-
-
-@pytest.fixture()
-def mocked_client():
-    with responses.RequestsMock() as rsps:
-        rsps.get('http://localhost:80/health', json={'status': 'ok'})
-        yield rsps
-
+from climatoology.utility.api import LulcWorkUnit, LulcUtility, LabelDescriptor
 
 unit_a = LulcWorkUnit(
     area_coords=(8.0859375, 47.5172006978394, 8.26171875, 47.63578359086485),
-    start_date='2023-05-01',
-    end_date='2023-06-01',
+    start_date=datetime.date(2023, 5, 1),
+    end_date=datetime.date(2023, 6, 1),
 )
-
 unit_b = LulcWorkUnit(
     area_coords=(8.0859375, 47.63578359086485, 8.26171875, 47.754097979680026),
-    start_date='2023-05-01',
-    end_date='2023-06-01',
+    start_date=datetime.date(2023, 5, 1),
+    end_date=datetime.date(2023, 6, 1),
 )
 
 
@@ -86,3 +77,32 @@ def test_lulc_when_encountering_connection_issues(mocked_client):
     with pytest.raises(PlatformUtilityException):
         with operator.compute_raster([unit_a]):
             pass
+
+
+def test_legend_retrieval(mocked_client):
+    mocked_client.get(
+        'http://localhost:80/segment/describe',
+        json={
+            'class1': {
+                'name': 'class1',
+                'description': 'description',
+                'osm_filter': 'landuse=nothing',
+                'raster_value': 1,
+                'color': [255, 255, 255],
+            }
+        },
+    )
+    operator = LulcUtility(host='localhost', port=80, path='/')
+
+    expected_result = {
+        'class1': LabelDescriptor(
+            name='class1',
+            description='description',
+            osm_filter='landuse=nothing',
+            raster_value=1,
+            color=(255, 255, 255),
+        )
+    }
+    computed_result = operator.get_class_legend()
+
+    assert expected_result == computed_result